Ft0

+++ src/app/elements/data-cards/data-cards.component.ts
@@ -1,8 +1,13 @@
 
-import { Component, Input, OnInit  } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output  } from '@angular/core';
+import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
+import { LoginsComponent } from 'src/app/logins/logins.component';
 import { UsuarioComponent } from 'src/app/logins/usuario/usuario.component';
 import { ReservasService } from 'src/app/services/reservas.service';
 import { UsuariosService } from 'src/app/services/usuarios.service';
+import { AuthService } from 'src/app/_services/auth.service';
+import { TokenStorageService } from 'src/app/_services/token-storage.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-data-cards',
@@ -14,7 +19,10 @@ export class DataCardsComponent implements OnInit {
     @Input() isReserva: any;
     @Input() reservaId: any;
 
-    constructor(private reservaService: ReservasService, private usuarioService: UsuariosService){}
+    @Output() cargarReservas = new EventEmitter<any>();
+
+    constructor(private reservaService: ReservasService, public usuarioService: UsuariosService, public tokenService: TokenStorageService,
+      public dialog: MatDialog,){}
 
 
     ngOnInit(): void {
@@ -41,7 +49,12 @@ export class DataCardsComponent implements OnInit {
         }
         console.log(reserva);
         this.reservaService.hacerReserva(reserva).subscribe(data =>{
-          console.log(data);
+          Swal.fire({
+            icon: 'success',
+            title: 'Reserva completada correctamente',
+            showConfirmButton: false,
+            timer: 1500
+          })
 
         });
       });
@@ -51,7 +64,45 @@ export class DataCardsComponent implements OnInit {
     }
 
     eliminarReserva(){
-      this.reservaService.eliminarReserva(this.reservaId).subscribe(data=>{
+
+
+
+      Swal.fire({
+        title: '¿Quieres eliminar la reserva?',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        cancelButtonText: 'cancelar',
+        confirmButtonText: 'Eliminar el texto'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          Swal.fire(
+            'actualizado!',
+            'Su perfil ha sido actualizado',
+            'success'
+          )
+          this.reservaService.eliminarReserva(this.reservaId).subscribe(data=>{
+            this.cargarReservas.emit();
+          })
+
+        }
+
+
       })
     }
+
+    openDialog(): void {
+
+      const dialogConfig = new MatDialogConfig();
+
+      dialogConfig.disableClose = true;
+      dialogConfig.autoFocus = true;
+      dialogConfig.width = "100%"
+      dialogConfig.enterAnimationDuration = "400ms";
+      dialogConfig.exitAnimationDuration = "1200ms";
+
+      this.dialog.open(LoginsComponent);
+    }
+
 }
